feat: new collection is now a modal
