+++ packages/outstatic/package.json
@@ -173,8 +173,8 @@
     "@types/lowlight": "^0.0.3",
     "@types/node": "^22.1.0",
     "@types/pluralize": "^0.0.29",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@types/testing-library__jest-dom": "^5.14.5",
     "concurrently": "^7.6.0",
     "eslint": "^8.57.0",
@@ -189,8 +189,8 @@
     "next-router-mock": "^0.9.7",
     "node-mocks-http": "^1.11.0",
     "postcss-import": "^15.1.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-test-renderer": "^18.2.0",
     "tailwindcss": "3.4.7",
     "ts-jest": "^29.1.2",
