+++ packages/outstatic/src/client/pages/collections/collections.tsx
@@ -10,6 +10,7 @@ import useOutstatic from '@/utils/hooks/useOutstatic'
 import { Settings, Trash } from 'lucide-react'
 import CollectionOnboarding from './_components/collection-onboarding'
 import LineBackground from '@/components/ui/outstatic/line-background'
+import NewCollectionModal from './_components/new-collection-modal'
 
 export default function Collections() {
   const { data: collections, isPending } = useCollections()
@@ -18,7 +19,7 @@ export default function Collections() {
   const [selectedCollection, setSelectedCollection] =
     useState<CollectionType | null>(null)
   const [showDeleteModal, setShowDeleteModal] = useState(false)
-
+  const [showNewCollectionModal, setShowNewCollectionModal] = useState(false)
   if (isPending) return <AdminLoading />
 
   return (
@@ -31,10 +32,8 @@ export default function Collections() {
         <>
           <div className="mb-8 flex h-12 items-center">
             <h1 className="mr-12 text-2xl">Collections</h1>
-            <Button asChild>
-              <Link href={`${dashboardRoute}/collections/new`}>
-                New Collection
-              </Link>
+            <Button onClick={() => setShowNewCollectionModal(true)}>
+              New Collection
             </Button>
           </div>
           <div className="max-w-5xl w-full grid md:grid-cols-3 gap-6">
@@ -87,6 +86,12 @@ export default function Collections() {
           collection={selectedCollection}
         />
       ) : null}
+      {showNewCollectionModal && (
+        <NewCollectionModal
+          open={showNewCollectionModal}
+          onOpenChange={setShowNewCollectionModal}
+        />
+      )}
     </AdminLayout>
   )
 }
