--- packages/outstatic/src/client/pages/new-collection/index.tsx
@@ -1,2 +1,0 @@
-import NewCollection from './new-collection'
-export default NewCollection
